+++ tests/test_pixel_coord.py
@@ -2,6 +2,7 @@ import math
 import os
 import tempfile
 
+import numpy as np
 import pytest
 
 import yirgacheffe as yg
@@ -84,19 +85,20 @@ def test_latlng_for_pixel(
     assert math.isclose(result[1], expected[1])
 
 def test_latlng_for_pixel_on_operator() -> None:
-    layer1 = YirgacheffeLayer(
-        Area(-10, 10, 10, -10),
-        MapProjection("epsg:4326", 0.2, -0.2),
-    )
-    layer2 = YirgacheffeLayer(
-        Area(-10, 10, 10, -10),
-        MapProjection("epsg:4326", 0.2, -0.2),
-    )
-    calc = layer1 + layer2
-    result = calc.latlng_for_pixel(0, 0)
-    expected = (10.0, -10.0)
-    assert math.isclose(result[0], expected[0])
-    assert math.isclose(result[1], expected[1])
+    data1 = np.zeros((20, 20))
+    data2 = np.ones((20, 20))
+    with (
+        yg.from_array(data1, (-10, 10), ("epsg:4326", (1.0, -1.0))) as layer1,
+        yg.from_array(data2, (-10, 10), ("epsg:4326", (1.0, -1.0))) as layer2,
+    ):
+        assert layer1.area == Area(-10, 10, 10, -10)
+        assert layer2.area == Area(-10, 10, 10, -10)
+
+        calc = layer1 + layer2
+        result = calc.latlng_for_pixel(0, 0)
+        expected = (10.0, -10.0)
+        assert math.isclose(result[0], expected[0])
+        assert math.isclose(result[1], expected[1])
 
 @pytest.mark.parametrize(
     "area,projection,coord,expected",
@@ -135,19 +137,20 @@ def test_pixel_for_latlng(
     assert result == expected
 
 def test_pixel_for_latlng_on_operator() -> None:
-    layer1 = YirgacheffeLayer(
-        Area(-10, 10, 10, -10),
-        MapProjection("epsg:4326", 0.2, -0.2),
-    )
-    layer2 = YirgacheffeLayer(
-        Area(-10, 10, 10, -10),
-        MapProjection("epsg:4326", 0.2, -0.2),
-    )
-    calc = layer1 + layer2
-    result = calc.pixel_for_latlng(10, -10)
-    expected = (0, 0)
-    assert math.isclose(result[0], expected[0])
-    assert math.isclose(result[1], expected[1])
+    data1 = np.zeros((20, 20))
+    data2 = np.ones((20, 20))
+    with (
+        yg.from_array(data1, (-10, 10), ("epsg:4326", (1.0, -1.0))) as layer1,
+        yg.from_array(data2, (-10, 10), ("epsg:4326", (1.0, -1.0))) as layer2,
+    ):
+        assert layer1.area == Area(-10, 10, 10, -10)
+        assert layer2.area == Area(-10, 10, 10, -10)
+
+        calc = layer1 + layer2
+        result = calc.pixel_for_latlng(10, -10)
+        expected = (0, 0)
+        assert math.isclose(result[0], expected[0])
+        assert math.isclose(result[1], expected[1])
 
 @pytest.mark.parametrize(
     "area,window,projection,pixel,expected",